breaking: Update package structure
