+++ .changeset/little-teeth-brake.md
@@ -1,0 +1,6 @@
+---
+"errf": minor
+---
+
+breaking: Update package structure
+
