+++ package.json
@@ -19,19 +19,12 @@
 		"dist"
 	],
 	"exports": {
-		".": "./dist/index.mjs",
-		"./utils": "./dist/utils.mjs"
+		".": "./dist/index.mjs"
 	},
 	"main": "./dist/index.mjs",
 	"module": "./dist/index.mjs",
 	"types": "./dist/index.d.mts",
-	"typesVersions": {
-		"*": {
-			"utils": [
-				"./dist/utils.d.mts"
-			]
-		}
-	},
+	"typesVersions": {},
 	"scripts": {
 		"start": "unbuild --silent && node ./dist/index.js",
 		"build": "unbuild",
